Modify project for local testing
Gretty is a gradle plugin which automatically handles SSL certificates
for HTTPS, among other conveniences.  Unfortunately, it seems to
conflict with the fatJar task.

In the DbConnection class, the web app was unable to find the CA certs
for Amazon, because the path had was not relative to the class
resources.  It's surprising this ever worked on Lambda.  Perhaps they
already have their own CA in their trust store.  Adding the correct
path lookup code resolved this issue for local testing.

The web.xml format seemed to require a class in a particular format,
which the Amazon SpeechletServlet does not quite accomodate.  This
required a small wrapper class to perform the necessary initialization
within the constructor.

For now this serves as a starting point that future work can extend
to be more conveniently integrated with the build and test process.
Until then, a possible workflow could be to create a branch based on
this one and merge changes from another feature branch, then test.